pyproject.toml
[project]
name = "persona"
version = "0.0.0"
description = "Open data contract POC for hosting API and MCP services"
authors = [
  {name="Jasper Ginn", email="jasperginn@gmail.com"}
]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fsspec>=2025.10.0",
    "httpx>=0.28.1",
    "orjson>=3.11.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
    "httpx>=0.27.0",
]

[project.scripts]
persona = "persona.cli:entrypoint"

[tool.ruff]
lint.ignore = ["E501"]
extend-exclude = [
  "__pycache__",
  ".eggs",
  ".git",
  ".venv",
  "build",
  "dist",
  "notebooks",
  ".cache"
]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.pyright]
exclude = [".venv", ".github", "docs", "tests", ".devcontainer"]
include = ["src", "scripts"]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
markers = [
  "integration: marks integration tests",
  "llm: marks tests that require LLM calls",
]
cache_dir = "/home/vscode/workspace/.cache/pytest"

[dependency-groups]
dev = [
    "nest-asyncio>=1.6.0",
    "pre-commit>=3.8.0",
    "ipykernel>=6.29.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-httpx>=0.27.0",
]
mcp = [
    "aiofiles>=25.1.0",
    "fastmcp>=2.13.0.2",
    "httpx>=0.28.1",
    "watchfiles>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
